# WebAPI/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Web API ",
      "description": "API del sistema de capacitación",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
      "http",
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
        "/api/getLogrosEmpleado/{idEmpleado}": {
            "get": {
                "tags": [
                    "LogroEmpleado"
                ],
                "summary": "Obtiene todos los logros de un empleado mediante su id",
                "description": "Regresa todos los logros de un empleado mediante la solicitud de su id",
                "operationId": "getLogrosEmpleado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                          "name": "idEmpleado",
                          "description": "Id del empleado a obtener",
                          "required": true,
                          "type": "integer",
                          "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LogroEmpleado"
                        }
                    }
                }
            }
        },
        "/api/getLogroEmpleado/{idLogro}/{idEmpleado}": {
            "get": {
                "tags": [
                    "LogroEmpleado"
                ],
                "summary": "Encuentra el logro de un empleado mediante el id de ambos",
                "description": "Regresa el logro de un empleado mediante el id solicitado de ambos",
                "operationId": "getLogroEmpleado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idLogro",
                        "description": "Id del logro a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                          "name": "idEmpleado",
                          "description": "Id del empleado a obtener",
                          "required": true,
                          "type": "integer",
                          "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LogroEmpleado"
                        }
                    }
                }
            }
        },
        "/api/addLogroEmpleado": {
            "post": {
                "tags": [
                    "LogroEmpleado"
                ],
                "summary": "Agrega un nuevo logro a un empleado",
                "description": "Agrega un nuevo logro a un empleado en la base de datos",
                "operationId": "addLogroEmpleado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo logro de un empleado",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LogroEmpleado"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LogroEmpleado"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateLogroEmpleado/{idLogro}/{idEmpleado}": {
            "put": {
                "tags": [
                    "LogroEmpleado"
                ],
                "summary": "Actualiza el logro de un empleado",
                "description": "Actualiza el logro de un empleado en la base de datos",
                "operationId": "updateLogroEmpleado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "idLogro",
                        "description": "Id del logro a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "idEmpleado",
                        "description": "Id del empleado a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevos del logro del empleado",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LogroEmpleado"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LogroEmpleado"
                        }
                    },
                    "400": {
                        "description": "Logro de empleado Inválido"
                    },
                    "404": {
                        "description": "Logro de empleado no encontrado"
                    }

                }
            }
        },
        "/api/deleteLogroEmpleado/{idLogro}/{idEmpleado}": {
            "delete": {
                "tags": [
                    "LogroEmpleado"
                ],
                "summary": "Borra el logro de un empleado",
                "description": "Borra el logro de un empleado de la base de datos",
                "operationId": "deleteLogroEmpleado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idLogro",
                        "description": "Id del logro a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "idEmpleado",
                        "description": "Id del empleado a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LogroEmpleado"
                        }
                    },
                    "400": {
                        "description": "Logro de empleado Inválido"
                    },
                    "404": {
                        "description": "Logro de empleado no encontrado"
                    }
                }
            }
        },
...
      "/api/getMaterialSemaforo/{idMaterialSemaforo}": {
          "get": {
              "tags": [
                  "MaterialSemaforo"
              ],
              "summary": "Encuentra un material de un semáforo por id",
              "description": "Regresa el material de un semáforo con el id solicitado",
              "operationId": "getMaterialSemaforo",
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                    "in": "path",
                      "name": "idMaterialSemaforo",
                      "description": "Id del material de un semáforo a consultar",
                      "required": true,
                      "type": "integer",
                      "format": "int64"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Operación exitosa",
                      "schema": {
                          "$ref": "#/definitions/MaterialSemaforo"
                      }
                  }
              }
          }
      },
      "/api/addMaterialSemaforo": {
          "post": {
              "tags": [
                  "MaterialSemaforo"
              ],
              "summary": "Agrega un nuevo material de un semáforo",
              "description": "Agrega un material de un semáforo a la base de datos",
              "operationId": "addMaterialSemaforo",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                      "in": "body",
                      "name": "body",
                      "description": "Nueva material de un semáforo",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/MaterialSemaforo"
                      }                      
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Operación exitosa",
                      "schema": {
                          "$ref": "#/definitions/MaterialSemaforo"
                      }
                  },
                  "405": {
                      "description": "Datos inválidos"
                  }
              }
          }
      },
      "/api/updateMaterialSemaforo/{idMaterialSemaforo}": {
          "put": {
              "tags": [
                  "MaterialSemaforo"
              ],
              "summary": "Actualiza un material de un semáforo",
              "description": "Actualiza un material de un semáforo en la base de datos",
              "operationId": "updateMaterialSemaforo",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                      "in": "path",
                      "name": "idMaterialSemaforo",
                      "description": "Id del material de un semáforo a actualizar",
                      "required": true,
                      "type": "integer",
                      "format": "int64"
                  },
                  {
                      "in": "body",
                      "name": "body",
                      "description": "Datos nuevos de un material de un semáforo",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/MaterialSemaforo"
                      }                      
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Operación exitosa",
                      "schema": {
                          "$ref": "#/definitions/MaterialSemaforo"
                      }
                  },
                  "400": {
                      "description": "Material de semáforo Inválida"
                  },
                  "404": {
                      "description": "Material de semáforo no encontrado"
                  }

              }
          }
      },
      "/api/deleteMaterialSemaforo/{idMaterialSemaforo}": {
          "delete": {
              "tags": [
                  "MaterialSemaforo"
              ],
              "summary": "Borra un material de un semáforo",
              "description": "Inactiva un material de un semáforo de la base de datos",
              "operationId": "deleteMaterialSemaforo",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                    "in": "path",
                      "name": "idMaterialSemaforo",
                      "description": "Id del material de un semáforo a borrar",
                      "required": true,
                      "type": "integer",
                      "format": "int64"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Operación exitosa",
                      "schema": {
                          "$ref": "#/definitions/MaterialSemaforo"
                      }
                  },
                  "400": {
                      "description": "Material de semáforo inválido"
                  },
                  "404": {
                      "description": "Material de semáforo no encontrado"
                  }
              }
          }
      }
    },
    "definitions": {
...
        "LogroEmpleado": {
            "type": "object",
            "properties": {
                "idEmpleado": {
                    "type": "integer",
                    "format": "int64"
                },
                "idLogro": {
                    "type": "integer",
                    "format": "int64"
                },
                "fechaObtencion": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Fecha y hora en que el empleado obtuvo un logro",
                    "example": "2021-07-21T17:32:28Z"
                }
            }
        },
        "MaterialSemaforo": {
          "type": "object",
          "properties": {
            "idMaterialSemaforo": {
              "type": "integer",
              "format": "int64"
            },
            "descripcion": {
              "type": "string"
            },
            "color": {
              "type": "string"
            }
          }
        }
    }
  }